+++ packages/cli/src/commands/pre/index.test.ts
@@ -63,9 +63,9 @@ describe("exitPre", () => {
     );
     await pre(cwd, { command: "exit" });
 
-    expect(await fs.readJson(path.join(cwd, ".changeset", "pre.json"))).toEqual(
-      { ...preStateForSimpleProject, mode: "exit" }
-    );
+    expect(
+      await fs.readJson(path.join(cwd, ".changeset", "pre.json"))
+    ).toEqual({ ...preStateForSimpleProject, mode: "exit" });
   });
   it("should throw if not in pre", async () => {
     let cwd = f.copy("simple-project");