+++ packages/cli/src/commands/version/version.test.ts
@@ -194,6 +194,41 @@ describe("running version in a simple project", () => {
     );
   });
 
+  it("should not commit the result if commit config is not set", async () => {
+    await writeChangesets([simpleChangeset2], cwd);
+    const spy = jest.spyOn(git, "commit");
+
+    expect(spy).not.toHaveBeenCalled();
+
+    await versionCommand(cwd, defaultOptions, modifiedDefaultConfig);
+
+    expect(spy).not.toHaveBeenCalled();
+  });
+
+  it("should commit the result if commit config is set", async () => {
+    await writeChangesets([simpleChangeset2], cwd);
+    const spy = jest.spyOn(git, "commit");
+
+    expect(spy).not.toHaveBeenCalled();
+
+    await versionCommand(cwd, defaultOptions, {
+      ...modifiedDefaultConfig,
+      commit: true
+    });
+
+    expect(spy).toHaveBeenCalled();
+    expect(spy.mock.calls[0][0]).toMatchInlineSnapshot(`
+      "RELEASING: Releasing 2 package(s)
+
+      Releases:
+        pkg-a@1.1.0
+        pkg-b@1.0.1
+
+      [skip ci]
+      "
+    `);
+  });
+
   describe("when there are multiple changeset commits", () => {
     it("should bump releasedPackages", async () => {
       await writeChangesets([simpleChangeset, simpleChangeset2], cwd);
@@ -321,6 +356,44 @@ describe("running version in a simple project with workspace range", () => {
   });
 });
 
+describe("same package in different dependency types", () => {
+  it("should update different range types correctly", async () => {
+    let cwd = f.copy("simple-project-same-dep-diff-range");
+    await writeChangeset(
+      {
+        releases: [
+          {
+            name: "pkg-b",
+            type: "patch"
+          }
+        ],
+        summary: "a very useful summary for the first change"
+      },
+      cwd
+    );
+
+    await version(cwd, defaultOptions, modifiedDefaultConfig);
+
+    let packages = (await getPackages(cwd))!;
+    expect(packages.packages.map(x => x.packageJson)).toEqual([
+      {
+        devDependencies: {
+          "pkg-b": "1.0.1"
+        },
+        peerDependencies: {
+          "pkg-b": "^1.0.1"
+        },
+        name: "pkg-a",
+        version: "1.0.0"
+      },
+      {
+        name: "pkg-b",
+        version: "1.0.1"
+      }
+    ]);
+  });
+});
+
 describe("snapshot release", () => {
   it("should update the packge to unique version no matter the kind of version bump it is", async () => {
     let cwd = f.copy("simple-project");
@@ -349,6 +422,27 @@ describe("snapshot release", () => {
         version: expect.stringContaining("0.0.0-exprimental-")
       })
     );
+  });
+
+  it("should not commit the result even if commit config is set", async () => {
+    let cwd = f.copy("simple-project");
+    await writeChangesets([simpleChangeset2], cwd);
+    const spy = jest.spyOn(git, "commit");
+
+    expect(spy).not.toHaveBeenCalled();
+
+    await versionCommand(
+      cwd,
+      {
+        snapshot: "exprimental"
+      },
+      {
+        ...modifiedDefaultConfig,
+        commit: true
+      }
+    );
+
+    expect(spy).not.toHaveBeenCalled();
   });
 
   describe("useCalculatedVersionForSnapshots: true", () => {
@@ -782,7 +876,7 @@ describe("pre", () => {
       }
     ]);
   });
-  it("should use the highest bump type for between all prereleases for every prerelease", async () => {
+  it("should use the highest bump type between all prereleases when versioning a package", async () => {
     let cwd = f.copy("simple-project");
     await writeChangeset(
       {
@@ -827,6 +921,114 @@ describe("pre", () => {
       {
         name: "pkg-b",
         version: "1.0.0"
+      }
+    ]);
+  });
+  it("should use the highest bump type between all prereleases when versioning a dependant package", async () => {
+    let cwd = f.copy("simple-project");
+    await pre(cwd, { command: "enter", tag: "next" });
+
+    await writeChangeset(
+      {
+        releases: [{ name: "pkg-a", type: "major" }],
+        summary: "a very useful summary for the first change"
+      },
+      cwd
+    );
+
+    await version(cwd, defaultOptions, modifiedDefaultConfig);
+    let packages = (await getPackages(cwd))!;
+
+    expect(packages.packages.map(x => x.packageJson)).toEqual([
+      {
+        dependencies: { "pkg-b": "1.0.0" },
+        name: "pkg-a",
+        version: "2.0.0-next.0"
+      },
+      {
+        name: "pkg-b",
+        version: "1.0.0"
+      }
+    ]);
+
+    await writeChangeset(
+      {
+        releases: [{ name: "pkg-b", type: "minor" }],
+        summary: "a very useful summary for the second change"
+      },
+      cwd
+    );
+    await version(cwd, defaultOptions, modifiedDefaultConfig);
+    packages = (await getPackages(cwd))!;
+
+    expect(packages.packages.map(x => x.packageJson)).toEqual([
+      {
+        name: "pkg-a",
+        version: "2.0.0-next.1",
+        dependencies: { "pkg-b": "1.1.0-next.0" }
+      },
+      {
+        name: "pkg-b",
+        version: "1.1.0-next.0"
+      }
+    ]);
+  });
+  it("should use the highest bump type between all prereleases for a linked package when versioning a dependant package", async () => {
+    let linkedConfig = {
+      ...modifiedDefaultConfig,
+      linked: [["pkg-a", "pkg-b"]]
+    };
+    let cwd = f.copy("linked-and-not-linked-packages");
+    await pre(cwd, { command: "enter", tag: "next" });
+    await writeChangeset(
+      {
+        releases: [{ name: "pkg-a", type: "minor" }],
+        summary: "a very useful summary"
+      },
+      cwd
+    );
+    await version(cwd, defaultOptions, linkedConfig);
+    let packages = (await getPackages(cwd))!;
+
+    expect(packages.packages.map(x => x.packageJson)).toEqual([
+      {
+        name: "pkg-a",
+        version: "1.1.0-next.0"
+      },
+      {
+        name: "pkg-b",
+        version: "1.0.0",
+        dependencies: { "pkg-c": "0.1.0" }
+      },
+      {
+        name: "pkg-c",
+        version: "0.1.0"
+      }
+    ]);
+
+    await writeChangeset(
+      {
+        releases: [{ name: "pkg-c", type: "patch" }],
+        summary: "a very useful summary"
+      },
+      cwd
+    );
+    await version(cwd, defaultOptions, linkedConfig);
+    packages = (await getPackages(cwd))!;
+
+    expect(packages.packages.map(x => x.packageJson)).toEqual([
+      {
+        name: "pkg-a",
+        version: "1.1.0-next.0"
+      },
+      {
+        name: "pkg-b",
+        version: "1.1.0-next.1",
+        dependencies: { "pkg-c": "0.1.1-next.0" }
+      },
+      {
+        name: "pkg-c",
+        version: "0.1.1-next.0"
       }
     ]);
   });